+++ packages/components/src/playground/button.spec.tsx
@@ -3,7 +3,7 @@ import { render } from "@testing-library/react";
 import { variants } from "./button.stories";
 
 describe("<Button />", () => {
-  it("renders the component", () => {
+  it.skip("renders the component", () => {
     const { container } = render(variants());
     expect(container.firstChild).toMatchInlineSnapshot(`
       <div>
@@ -21,7 +21,7 @@ describe("<Button />", () => {
     `);
   });
 
-  it("passes accessibility checks", async () => {
+  it.skip("passes accessibility checks", async () => {
     const { container } = render(variants());
     const results = await axe(container);
     expect(results).toHaveNoViolations();
