+++ packages/components/src/playground/button.stories.tsx
@@ -4,7 +4,7 @@ import Button from "./button";
 import { action } from "@storybook/addon-actions";
 
 export default {
-  title: "Button",
+  title: "ButtonPlayground",
   component: Button,
 };
